requirement.txt
appdirs==1.4.3
asgiref==3.2.3
Babel==2.6.0
certifi==2019.11.28
#cffi==1.14.0
chardet==3.0.4
cryptography==2.8
decorator==4.3.0
defusedxml==0.6.0
diff-match-patch==20181111
distlib==0.3.0
Django==2.1.7
django-appconf==1.0.3
django-cors-headers==3.2.1
django-imagekit==4.0.2
django-import-export==2.0.1
django-mathfilters==0.4.0
django-model-utils==4.0.0
django-thumbnail-maker==0.0.6
djangorestframework==3.11.0
djangorestframework-datatables==0.5.1
docutils==0.14
easy-thumbnails==2.7
ebaysdk==2.1.5
et-xmlfile==1.0.1
#feedparser==5.2.1
filelock==3.0.12
gdata==2.0.18
#gevent==1.1.2
#greenlet==0.4.10
html2text==2018.1.9
idna==2.8
importlib-metadata==1.5.0
importlib-resources==1.0.2
jdcal==1.4.1
jedi==0.15.2
Jinja2==2.10.1
libsass==0.17.0
#lxml==4.5.0
Mako==1.1.1
MarkupPy==1.14
MarkupSafe==1.1.0
mock==4.0.1
mysqlclient==1.4.6
ndg-httpsclient==0.5.1
ninja2==0.1
#numpy==1.18.1
odfpy==1.4.1
openpyxl==3.0.3
#pandas==0.25.3
parso==0.6.2
passlib==1.7.2
pbr==5.4.4
pilkit==2.0
#Pillow==7.0.0
pluggy==0.13.1
psutil==5.7.0
psycogreen==1.0.1
#psycopg2==2.8.4
pyasn1==0.4.8
pycparser==2.19
pydot==1.4.1
#pyOpenSSL==19.1.0
pyparsing==2.4.6
PyPDF2==1.26.0
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.9
python-openid==2.2.5
pytz==2019.1
pyusb==1.0.2
PyYAML==5.3
reportlab==3.5.34
requests==2.22.0
six==1.14.0
sorl-thumbnail==12.6.2
sqlparse==0.3.0
stevedore==1.32.0
stripe==2.42.0
#suds-jurko==0.6
tablib==0.14.0
ujson==1.35
urllib3==1.25.8
virtualenv==20.0.4
virtualenv-clone==0.5.3
#virtualenvwrapper==4.8.4
vobject==0.9.6.1
Werkzeug==1.0.0
xlrd==1.2.0
XlsxWriter==1.2.7
xlwt==1.3.0
zipp==3.0.0
